requirements.txt
beautifulsoup4==4.12.3
IMAPClient==3.0.1
pandas==2.2.3
python-dotenv==1.0.1
setuptools==70.0.0
lxml==5.3.0
openpyxl==3.1.5